# api/docs/swagger.yaml
definitions:
...
  models.Credentials:
    properties:
      password:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      username:
        type: string
    type: object
...
  models.Order:
    type: object
  models.Role:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-comments:
      Admin: "4"
      Customer: "0"
      Kitchen: "2"
      Manager: "3"
      Waiter: "1"
    x-enum-varnames:
    - Customer
    - Waiter
    - Kitchen
    - Manager
    - Admin
info:
  contact: {}
paths:
...
  /api/v1/login/credentials:
    patch:
      description: Updates user based on their credentials
      parameters:
      - description: user details
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Credentials'
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      summary: Updates a user
      tags:
      - user
  /api/v1/login/credentials/:
    post:
      description: Posts a user into the database
      parameters:
      - description: user details
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Credentials'
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      summary: Posts a user
      tags:
      - user
  /api/v1/login/credentials/{user}:
    delete:
      description: Deletes the user via their username
      parameters:
      - description: username
        in: path
        name: user
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      summary: Deletes a user
      tags:
      - user
  /api/v1/login/credentials/{username}:
    get:
      description: Returns the user based on their username
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      summary: Gets a user
      tags:
      - user
...
  /api/v1/orders:
    get:
      description: Returns a JSON object of all orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Get a list of all orders
      tags:
      - order
    patch:
      description: Updates the data of a currently existing order
      parameters:
      - description: JSON object of an order to be updated
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Patch an order
      tags:
      - order
    post:
      description: Adds an order if it does not already exist in the orders
      parameters:
      - description: JSON object of an order to be added
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Post an order
      tags:
      - order
    put:
      description: Puts an order, if it does not exist it is added otherwise it updates
        the order of the same ID
      parameters:
      - description: JSON object of an order to be added or updated
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
      summary: Put an order
      tags:
      - order
  /api/v1/orders/{id}:
    delete:
      description: Returns a status, depending on if the Order was deleted successfully
        or not
      parameters:
      - description: id of the order
        format: int
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete a specific order
      tags:
      - order
    get:
      description: Returns a JSON object of the order of the given ID
      parameters:
      - description: id of the order
        format: int
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Get a specific order
      tags:
      - order
  /api/v1/orders/{id}/{status}:
    patch:
      description: Returns the status of the given order ID
      parameters:
      - description: id of the order
        format: int
        in: path
        name: id
        required: true
        type: integer
      - description: status of the order
        format: int
        in: path
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Updates the status of a specific order
      tags:
      - order
  /api/v1/orders/{id}/status:
    get:
      description: Returns the status of the given order ID
      parameters:
      - description: id of the order
        format: int
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Get the status of a specific order
      tags:
      - order
...
  /api/v1/table/bills/{table}:
    get:
      description: Fetches the bill for specific table number and returns it.
      parameters:
      - description: table number
        format: int
        in: path
        name: table
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Gets bill for a specific table
      tags:
      - table
  /api/v1/table/orders/{table}:
    get:
      description: Fetches the orders for specific table number and returns them.
      parameters:
      - description: table number
        format: int
        in: path
        name: table
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Gets orders for a specific table
      tags:
      - table
...
swagger: "2.0"